version 1.0

+++ src/main/java/controller/admin_command/user/DeleteUserCommand.java
@@ -8,6 +8,12 @@ import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
 public class DeleteUserCommand implements Command {
+    /**
+     * removes user from database, admin can only remove passengers, main admin can remove passengers and admins
+     * @param request
+     * @param response
+     * @return
+     */
     @Override
     public String execute(HttpServletRequest request, HttpServletResponse response) {
         HttpSession session = request.getSession();
@@ -22,13 +28,12 @@ public class DeleteUserCommand implements Command {
             }
         }
 
-        if(adminRole == 1 && userToDeleteId == 1) {
+        if(adminRole == 1 && userToDeleteId != 1) {
+            AdminService.getInstance().deleteUser(userToDeleteId);
+        } else {
             request.setAttribute("cantDeleteUser", "cant delete user coz no rights");
         }
 
-        if(adminRole == 1) {
-            AdminService.getInstance().deleteUser(userToDeleteId);
-        }
         return "controller?action=getAllUsers";
     }
 }
